% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleAccumulate.R
\name{accumulate}
\alias{accumulate}
\title{accumulate toggle counts
returns accumulated counts (and only counts) for a (sub-hierarchy) of a parseResult (as returned by \link{parseToggles})}
\usage{
accumulate(sig, parseResult)
}
\arguments{
\item{sig}{the top signal of the sub-hierarchy for which counts shall be accumulated}

\item{parseResult}{what has been returned by \link{parseToggles}}
}
\description{
accumulate toggle counts
returns accumulated counts (and only counts) for a (sub-hierarchy) of a parseResult (as returned by \link{parseToggles})
}

