% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{buildParseReturn}
\alias{buildParseReturn}
\title{an interface to return parse results, setting sensible defaults}
\usage{
buildParseReturn(vcdfile = NA, timescale = NA, dumpstart = NA,
  hierarchy = NA, date = NA, version = NA, linesRead = 0)
}
\value{
An VCDFileObject containing information on \code{file}.
}
\description{
an interface to return parse results, setting sensible defaults
}
\keyword{internal}

