% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-data.tree.R
\name{FindFirstFast}
\alias{FindFirstFast}
\title{this expects a traversal and returns only the first match, behaves closely to data.tree::FindNode
trav <- Traverse(node, traversal, pruneFun = NULL, filterFun = NULL)}
\usage{
FindFirstFast(trav, name)
}
\arguments{
\item{trav}{a traversal generated by data.tree::Traverse}

\item{name}{the name of the node to search for}
}
\description{
this expects a traversal and returns only the first match, behaves closely to data.tree::FindNode
trav <- Traverse(node, traversal, pruneFun = NULL, filterFun = NULL)
}

