% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{FindNodeGeneric}
\alias{FindNodeGeneric}
\title{A more flexible variant of data.tree::FindNode(). This implementation allows to set a traversal and search by arbitrary fields.}
\usage{
FindNodeGeneric(node, value, field = "name", traversal = c("pre-order",
  "post-order", "in-order", "level", "ancestor"), all = F)
}
\arguments{
\item{node}{The root node of the tree to be searched.}

\item{value}{The value for which to seach.}

\item{field}{The field to be searched. The fieldname is given as a string.}

\item{traversal}{The order in which the tree should be traversed.}

\item{all}{whether only the first match or all matches shall be returned}
}
\value{
Depending on the value of `all` either a single node or a list of nodes. In the case that no match is found, `NULL` is returned.
}
\description{
`data.tree::Traverse` is used with the `traversal` argument to produce an ordering of the nodes which is subsequently searched. The `traversal` option therefore allows to exert some control over the order of results if the results are not unique.
}

