% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehelper.R
\name{nextLines}
\alias{nextLines}
\title{Read a bunch of lines from a file.}
\usage{
nextLines(con, skipLines = 0)
}
\arguments{
\item{con}{the connection (file) to read from}

\item{skipLines}{Skip this number of lines before reading data}
}
\value{
a list telling whether EOF has been reached and the data
}
\description{
Read a bunch of lines from a file.
}

