% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{noNA}
\alias{noNA}
\title{map NA entries to 0}
\usage{
noNA(x)
}
\arguments{
\item{x}{a (numerical) vector possibly containing NAs}
}
\value{
the input vector with NAs replaces by 0
}
\description{
map NA entries to 0
}
\keyword{internal}

