% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseBlock}
\alias{parseBlock}
\title{scope parsing dispatcher function to keep \code{parseVCDForKeys} readable
also eases handling of nested scopes (e.g. vars definitions)}
\usage{
parseBlock(tok, key)
}
\arguments{
\item{vcdfile}{The file to open.}
}
\value{
An list containing the parse results for \code{file}.
}
\description{
scope parsing dispatcher function to keep \code{parseVCDForKeys} readable
also eases handling of nested scopes (e.g. vars definitions)
}

