% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHierarchy.R
\name{parseHierarchy}
\alias{parseHierarchy}
\title{Parse a hierarchy from a VCD File}
\usage{
parseHierarchy(tok)
}
\arguments{
\item{tok}{a \link{Tokenizer} set up to read from the VCDFile}
}
\value{
returns the parsed hierarchy
}
\description{
This function parses the module/variable hierarchy from the VCDFile into a
\code{data.tree} structure.
}

