% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHierarchy.R
\name{parseScope}
\alias{parseScope}
\title{Parse a scope statement from a VCD File}
\usage{
parseScope(node, tok)
}
\arguments{
\item{node}{the hierarchy where nodes leaves shall be appended}

\item{tok}{a \link{Tokenizer} set up to read from the VCDFile}
}
\value{
returns the parsed hierarchy
}
\description{
This function parses a single scope from the VCDFile into a
\code{data.tree} (subtree-)structure.
}

