% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseToggles.R
\name{parseToggles}
\alias{parseToggles}
\title{Parses a VCD file and generated toggle count statistics.}
\usage{
parseToggles(vcd, top = NA, depth = 0L)
}
\arguments{
\item{vcd}{The VCDFile to parse}

\item{top}{The top signal, if none is given the top signal of the vcdfile is taken}

\item{depth}{depth in the module tree before signals are accumulated, default=0 (top only), -1 is infinite depth

For more elaborate filtering of nodes, pre-edit the signal-tree in the vcd-object.
Black/Whitelisting of signals is an optional further enhancement.}
}
\value{
a module tree and a list of toggle counts and a list of dump events

parseToggles(vcd,"top",3)
parseToggles(vcd,"SBOX1",3)
}
\description{
Parses a VCD file and generated toggle count statistics.
}

