% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseVCDForKeys}
\alias{parseVCDForKeys}
\title{Parse a VCDFile for a given list of sections
This function is the main workhorse, delegating to speciaclised
functions for parsing the individual fields.
Altogether this implements a top-down parser for VCD files.}
\usage{
parseVCDForKeys(vcdfile, keys, header)
}
\arguments{
\item{vcdfile}{The file to open.}
}
\value{
An list containing the parse results for \code{file}.
}
\description{
Parse a VCDFile for a given list of sections
This function is the main workhorse, delegating to speciaclised
functions for parsing the individual fields.
Altogether this implements a top-down parser for VCD files.
}

