% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseVCDHeader}
\alias{parseVCDHeader}
\title{Parse a VCDFile for its header fields}
\usage{
parseVCDHeader(vcdfile)
}
\arguments{
\item{vcdfile}{The file to open.}
}
\value{
An list containing the parse results for \code{file}.
}
\description{
Parse a VCDFile for its header fields
}
\keyword{internal}

