% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayVCD.R
\name{plotToggles}
\alias{plotToggles}
\title{the high level plotting function, vurrently provides plotting through ggplot2, dygraphs and plotly. The former being the fastes one, the others providing some degree of interactivity.}
\usage{
plotToggles(vcd, parse, top = NA, weights = list(`0` = -1, `1` = 1, z = 0, x
  = 1), type = c("ggplot2", "dygraphs", "plotly"), toggle_hold_time = 1,
  ...)
}
\arguments{
\item{vcd}{A vcd object containg a hierarchy, VCD file information,...}

\item{parse}{Parsed toggle events as generated by \link{parseToggles}}

\item{top}{The signal for which the toggle counts shall be plotted}

\item{weights}{Optional weights when toggles should be weighted, e.g. toggles to 0 should generate a smaller amplitude than toggles to 1-}

\item{type}{The type of plot, currently one of `c("ggplot2","dygraphs", "plotly")` with `ggplot2` being default.}

\item{toggle_hold_time}{Stretch a toggle event this amount of units into the future for better visibility. This is effectively the width of a toggle count in the plot; should match at most half a clock cycle, `vcd$timescale`` is a good referece value. Set to `NA`` to turn this feature off.}

\item{...}{other parameters that are forwarded to ploty or dygraph or a named vector of timestamps called events; these are printed as labels (currently in dygraphs only)}
}
\description{
the high level plotting function, vurrently provides plotting through ggplot2, dygraphs and plotly. The former being the fastes one, the others providing some degree of interactivity.
}

