% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcd2r.R
\name{VCDFile}
\alias{VCDFile}
\title{Create an VCDFileObject}
\usage{
VCDFile(filename = NA_character_, parseHeader = T)
}
\arguments{
\item{filename}{The file to open.}

\item{parseHeader}{A flag whether the header should be parsed on opening the file. Default is TRUE.}
}
\value{
An VCDFileObject containing information on \code{file}.
}
\description{
Create an VCDFileObject
}
\examples{
\dontrun{
VCDFile()
VCDFile("inter.vcd")
VCDFile("inter.vcd",F)
}
}

