% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKeywords.R
\name{parseBlock}
\alias{parseBlock}
\title{scope parsing dispatcher function to keep \code{parseVCDForKeys} readable
also eases handling of nested scopes (e.g. vars definitions)}
\usage{
parseBlock(tok, key)
}
\arguments{
\item{tok}{A tokenizer wrapping the VCD file to read from}

\item{key}{The keyword for dispatching}
}
\value{
An list containing the parse results for \code{file}.
}
\description{
scope parsing dispatcher function to keep \code{parseVCDForKeys} readable
also eases handling of nested scopes (e.g. vars definitions)
}
\keyword{internal}

